Separated everything out from MainActivity-Added location and NavDrawer classes
