+++ app/src/main/java/com/sparkmap/sparkmap/Location.java
@@ -1,0 +1,14 @@
+package com.sparkmap.sparkmap;
+
+import android.app.Activity;
+
+/**
+ * Created by Nate on 9/14/2017.
+ */
+
+public class Location{
+    Activity activity;
+    public Location(Activity activity) {
+        this.activity = activity;
+    }
+}
